+++ index.html
@@ -149,8 +149,8 @@
         >
           TRACKERS
         </button>
-        <a href="../html/overviewpage.html" class="nav-link">OVERVIEW</a>
-        <a href="../html/settingspage.html" class="nav-link">SETTINGS</a>
+        <a href="../src/html/overviewpage.html" class="nav-link">OVERVIEW</a>
+        <a href="../src/html/settingspage.html" class="nav-link">SETTINGS</a>
       </nav>
 
       <nav id="trackersMenuContainer">
@@ -159,9 +159,9 @@
           class="trackersMenu"
           aria-label="Select tracker type"
         >
-          <li><a href="../html/pomodoromode.html" class="menu-link">Study</a></li>
-          <li><a href="../html/trainingmode.html" class="menu-link">Exercise</a></li>
-          <li><a href="../index.html" class="menu-link">Work</a></li>
+          <li><a href="../src/html/pomodoromode.html" class="menu-link">Study</a></li>
+          <li><a href="../src/html/trainingmode.html" class="menu-link">Exercise</a></li>
+          <li><a href="../src/index.html" class="menu-link">Work</a></li>
         </ul>
       </nav>
     </header>
