+++ package.json
@@ -23,18 +23,18 @@
     "transform": {},
     "testEnvironment": "node",
     "collectCoverageFrom": [
-      "scripts/**/*.js",
-      "!scripts/timerDOM.js",
-      "!scripts/script.js",
-      "!scripts/navigation.js",
-      "!scripts/training.js",
-      "!scripts/settingsDOM.js",
-      "!scripts/barchartDOM.js",
-      "!scripts/localStorage.js",
-      "!scripts/overviewDOM.js",
-      "!scripts/statsDisplayDOM.js",
-      "!scripts/swCall.js",
-      "!scripts/addTimeDOM.js"
+      "src/scripts/**/*.js",
+      "!src/scripts/timerDOM.js",
+      "!src/scripts/timer.js",
+      "!src/scripts/navigation.js",
+      "!src/scripts/training.js",
+      "!src/scripts/settingsDOM.js",
+      "!src/scripts/barchartDOM.js",
+      "!src/scripts/localStorage.js",
+      "!src/scripts/overviewDOM.js",
+      "!src/scripts/statsDisplayDOM.js",
+      "!src/scripts/swCall.js",
+      "!src/scripts/addTimeDOM.js"
     ],
     "coverageThreshold": {
       "global": {
