+++ src/html/pomodoromode.html
@@ -5,14 +5,18 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta
       name="description"
-      content="Customize your experience with dark mode and manage your local storage for optimal performance."
+      content="Track your study sessions and view your progress with our Pomodoro timer."
     />
-    <title>Study Pomodoro - Settings</title>
+    <title>Study - Study Pomodoro</title>
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
       href="https://fonts.googleapis.com/css2?family=Abril+Fatface&family=Bricolage+Grotesque:opsz,wght@12..96,200..800&display=swap"
       rel="stylesheet"
+    />
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css"
     />
     <link rel="manifest" href="/manifest.json" />
     <meta name="theme-color" content="#673ab7" />
@@ -133,7 +137,7 @@
     <link rel="stylesheet" href="../css/components.css" />
     <link rel="stylesheet" href="../css/timer.css" />
   </head>
-  <body>
+  <body class="timerPage">
     <header>
       <nav class="headerContainer" aria-label="Main Navigation">
         <button
@@ -146,7 +150,7 @@
           TRACKERS
         </button>
         <a href="overviewpage.html" class="nav-link">OVERVIEW</a>
-        <a href="settingspage.html" class="nav-link active">SETTINGS</a>
+        <a href="settingspage.html" class="nav-link">SETTINGS</a>
       </nav>
 
       <nav id="trackersMenuContainer">
@@ -157,32 +161,126 @@
         >
           <li><a href="pomodoromode.html" class="menu-link">Study</a></li>
           <li><a href="trainingmode.html" class="menu-link">Exercise</a></li>
-          <li><a href="../index.html" class="menu-link">Work</a></li>
+          <li><a href="/index.html" class="menu-link">Work</a></li>
         </ul>
       </nav>
     </header>
 
     <main>
-      <header class="settingsHeaderContainer">
-        <h1 id="settingsHeaderArea">Settings</h1>
+      <header class="timerHeaderContainer">
+        <h1 id="timerHeading">Study</h1>
+        <p>Study mode</p>
       </header>
 
-      <section class="settingsChoices" aria-labelledby="settingsHeaderArea">
-        <h2 class="visually-hidden">Preferences and Data Management</h2>
+      <section class="timer-card" aria-labelledby="timerHeading">
+        <h2 class="visually-hidden">Study timer display</h2>
+        <time
+          class="timeDisplay"
+          id="timeDisplay"
+          role="timer"
+          aria-live="polite"
+        >
+          00:00
+        </time>
 
-        <button id="darkModeButton" class="btn-primary" type="button">
-          Dark Mode
-        </button>
+        <div class="mainControls">
+          <button id="playBtn" type="button" aria-label="Start Study Session">
+            <i class="fa-solid fa-play" aria-hidden="true"></i>
+          </button>
+          <button id="pauseBtn" type="button" aria-label="Pause Study Session">
+            <i class="fa-solid fa-pause" aria-hidden="true"></i>
+          </button>
+          <button id="stopBtn" type="button" aria-label="Stop Study Session">
+            <i class="fa-solid fa-stop" aria-hidden="true"></i>
+          </button>
+        </div>
+      </section>
 
-        <button id="localStorageResetButton" class="btn-primary" type="button">
-          Reset Local Storage
-        </button>
+      <section
+        class="statsContainer"
+        aria-labelledby="statsArea"
+        data-category="Study"
+      >
+        <h2 id="statsArea">This Week</h2>
+        <div class="stats-grid">
+          <div class="statsDisplay">
+            <time id="totalTime" class="number-card">0h</time>
+            <span>Total Time</span>
+          </div>
+
+          <div class="statsDisplay">
+            <time id="currentSession" class="number-card">0h</time>
+            <span>Current Session</span>
+          </div>
+
+          <div class="statsDisplay">
+            <time id="longestSession" class="number-card">0h</time>
+            <span>Longest Session</span>
+          </div>
+        </div>
+      </section>
+
+      <!--Extra time area-->
+
+      <section class="extraTimeContainer" aria-labelledby="extraTimeHeading">
+        <h2 id="extraTimeHeading" class="visually-hidden">Extra time area</h2>
+        <div class="extraTimeRow">
+          <div class="timeUnit">
+            <span class="dateLabel">Date</span>
+            <label class="dateIcon">
+              <i class="fa-regular fa-calendar"></i>
+              <input type="date" id="dateInput" aria-label="Select date" />
+            </label>
+          </div>
+          <div class="timeUnit">
+            <label for="timeInput">Start Time</label>
+            <input type="time" id="timeInput" aria-label="Select time" />
+          </div>
+          <div class="timeUnit">
+            <label for="hoursInput">Hours</label>
+            <input
+              type="number"
+              id="hoursInput"
+              inputmode="numeric"
+              pattern="[0-9]*"
+              min="0"
+              max="99"
+              value="0"
+            />
+          </div>
+          <div class="timeUnit">
+            <label for="minutesInput">Minutes</label>
+            <input
+              type="number"
+              id="minutesInput"
+              inputmode="numeric"
+              pattern="[0-9]*"
+              min="0"
+              max="59"
+              value="0"
+            />
+          </div>
+          <div class="timeUnit">
+            <label for="secondsInput">Seconds</label>
+            <input
+              type="number"
+              id="secondsInput"
+              inputmode="numeric"
+              pattern="[0-9]*"
+              min="0"
+              max="59"
+              value="0"
+            />
+          </div>
+          <button class="addTimeButton" id="addTimeButton">ADD TIME</button>
+        </div>
       </section>
     </main>
-
-    <script src="../scripts/script.js" type="module"></script>
-    <script src="../scripts/navigation.js"></script>
-    <script src="../scripts/settingsDOM.js" type="module"></script>
+    <script type="module" src="../scripts/navigation.js"></script>
+    <script type="module" src="../scripts/statsDisplayDOM.js"></script>
+    <script type="module" src="../scripts/timerDOM.js"></script>
+    <script type="module" src="../scripts/addTimeDOM.js"></script>
     <script src="/scripts/swCall.js"></script>
+    <script src="../scripts/settingsDOM.js" type="module"></script>
   </body>
 </html>
