+++ src/html/settingspage.html
@@ -5,9 +5,9 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <meta
       name="description"
-      content="Track your activities and view your progress with our comprehensive overview page. Access detailed history and insightful graphs."
+      content="Customize your experience with dark mode and manage your local storage for optimal performance."
     />
-    <title>Study Pomodoro - Overview</title>
+    <title>Study Pomodoro - Settings</title>
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
@@ -131,7 +131,7 @@
     <link rel="stylesheet" href="../css/variables.css" />
     <link rel="stylesheet" href="../css/base.css" />
     <link rel="stylesheet" href="../css/components.css" />
-    <link rel="stylesheet" href="../css/overview.css" />
+    <link rel="stylesheet" href="../css/timer.css" />
   </head>
   <body>
     <header>
@@ -141,11 +141,12 @@
           aria-label="Toggle tracker menu"
           aria-haspopup="true"
           aria-expanded="false"
+          aria-controls="trackersMenu"
         >
           TRACKERS
         </button>
-        <a href="overviewpage.html" class="nav-link active">OVERVIEW</a>
-        <a href="settingspage.html" class="nav-link">SETTINGS</a>
+        <a href="overviewpage.html" class="nav-link">OVERVIEW</a>
+        <a href="settingspage.html" class="nav-link active">SETTINGS</a>
       </nav>
 
       <nav id="trackersMenuContainer">
@@ -156,94 +157,32 @@
         >
           <li><a href="pomodoromode.html" class="menu-link">Study</a></li>
           <li><a href="trainingmode.html" class="menu-link">Exercise</a></li>
-          <li><a href="../index.html" class="menu-link">Work</a></li>
+          <li><a href="/index.html" class="menu-link">Work</a></li>
         </ul>
       </nav>
     </header>
 
     <main>
-      <section class="overviewHeaderContainer" aria-labelledby="mainTitle">
-        <h1 id="mainTitle">Overviews</h1>
-        <div class="view-toggle-group" role="tablist">
-          <button
-            id="historyButton"
-            class="btn-primary"
-            role="tab"
-            aria-selected="true"
-            aria-controls="historyOverviewArea"
-          >
-            History
-          </button>
-          <button
-            id="graphButton"
-            class="btn-primary"
-            role="tab"
-            aria-selected="false"
-            aria-controls="graphView"
-          >
-            Graph
-          </button>
-        </div>
-      </section>
+      <header class="settingsHeaderContainer">
+        <h1 id="settingsHeaderArea">Settings</h1>
+      </header>
 
-      <section
-        class="historyContainer"
-        id="historyOverviewArea"
-        role="tabpanel"
-        aria-labelledby="historyButton"
-      >
-        <h2 class="visually-hidden">Activity History</h2>
-        <ul class="historyList"></ul>
-      </section>
+      <section class="settingsChoices" aria-labelledby="settingsHeaderArea">
+        <h2 class="visually-hidden">Preferences and Data Management</h2>
 
-      <section
-        class="graphContainer"
-        id="graphView"
-        role="tabpanel"
-        aria-labelledby="graphButton"
-        hidden
-      >
-        <h2 class="visually-hidden">Activity Graph</h2>
-        <ul class="activityList">
-          <li class="activity graphItem" data-activity="work">
-            <div class="graphInfo">
-              <p class="graphCategory">WORK</p>
-            </div>
-            <time class="graphDuration" datetime="PT0H">0h</time>
-          </li>
-
-          <li class="activity graphItem" data-activity="study">
-            <div class="graphInfo">
-              <p class="graphCategory">STUDY</p>
-            </div>
-            <time class="graphDuration" datetime="PT0H">0h</time>
-          </li>
-
-          <li class="activity graphItem" data-activity="exercise">
-            <div class="graphInfo">
-              <p class="graphCategory">EXERCISE</p>
-            </div>
-            <time class="graphDuration" datetime="PT0H">0h</time>
-          </li>
-        </ul>
-      </section>
-
-      <section class="overviewSelectors" aria-label="Data range filters">
-        <button id="rangeSelectorButton" class="btn-primary">
-          SELECT <br />
-          RANGE
+        <button id="darkModeButton" class="btn-primary" type="button">
+          Dark Mode
         </button>
-        <button id="allTimeSelectorButton" class="btn-primary">
-          RESET <br />
-          RANGE
+
+        <button id="localStorageResetButton" class="btn-primary" type="button">
+          Reset Local Storage
         </button>
       </section>
     </main>
 
+    <script src="../scripts/timer.js" type="module"></script>
     <script src="../scripts/navigation.js"></script>
-    <script src="../scripts/barchartDOM.js" type="module"></script>
-    <script src="../scripts/overviewDOM.js" type="module"></script>
-    <script src="/scripts/swCall.js"></script>
     <script src="../scripts/settingsDOM.js" type="module"></script>
+    <script src="/scripts/swCall.js"></script>
   </body>
 </html>
