+++ src/html/trainingmode.html
@@ -164,7 +164,7 @@
         >
           <li><a href="pomodoromode.html" class="menu-link">Study</a></li>
           <li><a href="trainingmode.html" class="menu-link">Exercise</a></li>
-          <li><a href="../index.html" class="menu-link">Work</a></li>
+          <li><a href="/index.html" class="menu-link">Work</a></li>
         </ul>
       </nav>
     </header>
