+++ src/images/icon-192.png
