+++ src/images/icon-512.png
