+++ src/scripts/addTimeLogic.js
