+++ src/scripts/barchartDOM.js
