+++ src/scripts/barchartLogic.js
